etcd.py: log a warning on empty responses from etcd
Avoid trying to load empty etcd responses as JSON, spamming the logs.

In particular, this currently happens every time the connection to etcd
is lost. Our EtcdClient inherits from twisted.web.http.HTTPClient, which
handles lost connections by calling handleResponse. In case of lost
connections, the response buffer might be empty, thus logging a
confusing JSON error.

If the response is not emtpy and not valid JSON, log it together with
the exception.

Bug: T134893
Change-Id: I0adf7e77a3b62c2bdbbcc7cc23f7e3eca08daa9e