Change the default LVS BGP behavior per service
Change the per LVS service BGP announcing behavior to yes. This means
that a service that does not specify 'bgp': 'yes' in the configuration
will be announced via BGP. That is a change in the default behavior but
it's warranted. Our usage shows that in low-traffic LVS servers 39/40
services do get announced via BGP. In high-traffic LVS servers it's 2/5
and 11/19 and that's due to preconception that we should set bgp to no when
reusing an LVS IP address for say serving HTTPS as well as HTTP or TCP
as well as UDP. However the code for BGPFailover.addPrefix() actually uses
a python set() for storing the prefixes meaning it's idempotent and it's
fine to add the same prefix multiple times

Cherry-picked from 1.13

Change-Id: I6fe986cfb4cc7b46edecc2a7fa28f7a18cdb5177