Create new BGP message classes for incremental construction
The existing construct* methods in the BGP class are simple
and work well, but are not really flexible or optimal for large
messages such as UPDATEs.

Introduce a new ancestor class BGPMessage and derived BGPUpdateMessage
for efficient and incremental package construction.

Also move BGP.constructHeader to BGPMessage.prependHeader as a static
method.

Change-Id: I542ca34c09de45005b083c3f5f7797653f664989