Set empty PYTHONPATH in tox.ini
Our jenkins instance is not actually running any pybal tests. The reason
is that we're using `trial pybal` in tox.ini and funnily enough trial
requires PYTHONPATH to be set (even if it is set to an empty string) or
else it won't manage to find any tests and just pretend everything is
fine, claiming the tests have passed.

Change-Id: I6541a8384bd8284bcc76adec04c6e2d2febecda1
Ref: http://stackoverflow.com/questions/42425901/twisted-trial-pythonpath-and-sys-path