coordinator: exit if exceptions are raised importing monitors
If importing a configured pybal.monitor results in an exception, we want
PyBal to stop immediately. This is to be explicit with the admin who
might think that everything is fine while in fact none of the checks
that should have been performed by the failing monitor are being run.

Also, use importlib instead of __import__.

Change-Id: I5a0733310d0c8277e755e4cd1e3f1a61bc716a54