Stop pybal on all failures to properly parse 'monitors'
When a monitor module failed to import, pybal was made to exit in order
to prevent it from silently doing nothing. However, a failure parsing
the 'monitors' configuration option would only log an error and still
not have the desired effect. Stop pybal in more cases when it fails to
interpret 'monitors'.

Also complete unit testing coverage of createMonitoringInstances

Change-Id: I804af7081f645f4d2e90499a2394c40712a52241