Introduce server.is_pooled and make server.pooled usage more consistent
The usage of, and therefore the meaning of server.pooled had gotten
inconsistent, sometimes used to indicate the intention to (immediately)
pool (before the fact), and sometimes to indicate the actual status
of whether the server had been pooled in IPVS (after the fact).
Clarify the use by introducing a second variable .is_pooled to denote
the latter (managed by IPVS), and restrict the use of .pooled to the
former (managed by Coordinator).

We could consider renaming .pooled to .pool (indicating intention).
This ties into metric names as well.

Change-Id: Ib11e8e9ed968cff3e9f17c3946ba19923f34474a