Fix testRepool test case for previously-down-but-pooled
The test was incorrect; repool should *depool*
previously-down-but-pooled servers if depool threshold allows,
not also pool other servers. Also, with .removeServer mocked,
.pooled is not set as a side effect and thus can't be relied
on for testing.

Test whether lvsservice.removeServer gets called with the right
argument, instead of relying on the ambiguous .pooled (which semantics
we're also changing slightly in subsequent commits).

Change-Id: Ie950e29fe250f60189628a50ee79d8ddc918eafe