etcd: Introduce reconnectTimeout
In production, we have observed an issue with pybal losing its
connections to etcd shortly after a bunch of unhandled
twisted.internet.error.ConnectionLost errors.

See https://phabricator.wikimedia.org/P6711 for an example of such
situation.

In a test environment, we have verified that pybal does not reconnect to
etcd in case of etcd restarts.

Introduce a 1 second timeout between reconnection attempts to be used in
case of lost connections or connection failures and log the reconnection
attempts themselves. This fixes the reconnection issue in a testing
environment when issuing a systemctl etcd restart as well as when
stopping etcd altogher and starting it again after a few seconds.

Bug: T169765
Change-Id: I0db3a64af405aa11aaf3a215a8d57a8857d48dbd