Fix handling of withdrawals for Inet Unicast
Sending withdrawals for Inet Unicast was entirely broken due to 2
issues:

Advertisements were not being popped from the set 'withdrawals' as they
were taken from a prefix set 'withdrawalPrefixSet' instead, thereby
looping over it indefinitely.

The code did not handle the case of withdrawals when there are no
UPDATEs and would skip sending withdrawals entirely if attributeMap was
empty. When there are no UPDATEs, create a entry with empty attributes
so withdrawals are processed.

Change-Id: I889a54ad0c2ed252a12d68596785f2bb356d2d83