Restore testing of the UDP monitor check method
Previously the monitor invoked LoopingCall.start with now=True,
immediately invoking the check() call which was subsequently
tested as part of testing the run() method. Now, consistent with
the rest of the Pybal monitors, LoopingCall.start is invoked with
(now=False) by the ancestor class, which implies that the check()
method is not immediately invoked, and therefore wasn't tested
anymore.

Change-Id: I36720547fdcd223b003908c15feb4933ec070f55