Don't recalculate server.up in refreshPreexistingServer
refreshPreexistingServer should only be called for servers that
previously existed, not servers that have just been created and have
not yet been initialized. Consequently they already have monitors
setup and running, and server.up (the AND quantification of all
monitor up states) should already be accurate. No need to recalculate.

For now, let's assert that this is indeed the case.

Change-Id: Ibcc3d59e1e576ab5fe41acdcc4c36bea8ba24d8f