Remove Server.modified and refresh preexisting servers individually
Pybal's server state variable 'modified' was misnamed and therefor
very confusing. According to the implementation it really means
"preexisting, i.e. not new", as it's only set to False when a
Server instance is built, and kept at True for every server config
update thereafter, whether the server's configuration details are
changed or not.

Rather than rename the variable, remove it entirely as it's not really
needed. It was only used in Coordinator.refreshModifiedServers to
determine which servers should have their pool status recalculated.

Instead, simply refresh .pool status of preexisting servers during
the iteration of the new configuration instead of afterwards.

Rename Coordinator.refreshModifiedServers to .refreshPreexistingServer
accordingly, and have it accept a single servers to refresh.

Change-Id: Ia455ed63e0812942a34884fde3a9cad2337c7edc