Wait for onConfigUpdate initialization in setServers using inlineCallbacks
CoordinatorTestCase.setServers is a helper method that sets up
a test coordinator's server list, but it can be a bit misleading.
It calls onConfigUpdate (which isn async function by initializing
new servers).
For this, onConfigUpdate needs to return the DeferredList.

Also, any server parameters passed to setServers are set but NOT passed to
onConfigUpdate. Clarify this in the docstring.

Change-Id: I0566667262446b98d1482f59c9af8512acef186a