Expand Coordinator.resultUp behavior on first monitor check result
When initializing new Server instances, before any monitor results are
in, Pybal initializes server.up to a default value of True for the time
being. This is undesirable behavior because it's not easy to distinguish
states "up" from "Unknown" (or None), and should be rectified in the
FSM implementation in Pybal 2.

For now, treat resultUp differently when all monitors have completed
their first check and ensure Coordinator.repool is called in that case.
This ensures that any down-but-pooled servers can be depooled when
possible, e.g. when new servers have been added that are up.

This could alternatively be done immediately on new server list
initialization, but because at that point no actual monitor results are in
yet, it could cause even more churn.

Change-Id: I951194c6255418bfab0514683c926dcdfa86e1d1