Reduce reconnectTimeout for etcd to 0.1 seconds
The reconnect timeout was introduced when trying to solve the issue of
pybal not reconnecting to etcd when its reference etcd went offline
for a time. While we later found out the actual issue to be different
- basically that pybal would be requiring for an inexistent etcd index
in that situation - the reconnect timeout solved some race conditions
and was thus left in the mix.

Given we now need to do a coordinated rolling restart of the
application servers, I propose to reduce that number to 1 tenth of a
second, allowing us to reduce the minimum time required for a full
rolling restart to ~ 1 minute.

Bug: T264362
Change-Id: I177ebb8521d87f2fdbc64550645e490c143a3a2c